+++ package.json
@@ -4,6 +4,7 @@
     "@mantine/hooks": "^5.3.0",
     "@primer/octicons-react": "^17.5.0",
     "classnames": "^2.3.2",
-    "swr": "^1.3.0"
+    "swr": "^1.3.0",
+    "zustand": "^4.1.5"
   }
 }
